requirements.txt
streamlit
opencv-python
numpy
tensorflow
keras
cvzone
mediapipe
pyttsx3
pyenchant
Pillow
pymongo
python-dotenv